% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{as.data.frame.specr.object}
\alias{as.data.frame.specr.object}
\title{Return data.frame from specr.object}
\usage{
\method{as.data.frame}{specr.object}(x, ...)
}
\description{
Return data.frame from specr.object
}
\keyword{internal}
