% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{as.data.frame.specr.setup}
\alias{as.data.frame.specr.setup}
\title{Return tibble from specr.setup object}
\usage{
\method{as.data.frame}{specr.setup}(x, ...)
}
\description{
Return tibble from specr.setup object
}
\keyword{internal}
