% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{as_tibble.specr.object}
\alias{as_tibble.specr.object}
\title{Return tibble from specr.object}
\usage{
\method{as_tibble}{specr.object}(x, ...)
}
\description{
Return tibble from specr.object
}
\keyword{internal}
