% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{plot.specr.boot}
\alias{plot.specr.boot}
\title{Plot specification curve and under-the-null distributions}
\usage{
\method{plot}{specr.boot}(x, ...)
}
\arguments{
\item{x}{A \code{specr.boot} object, resulting from calling \code{boot_null()}.}

\item{...}{further arguments passed to or from other methods (currently ignored).}
}
\value{
A \link[ggplot2]{ggplot} object that can be customized further.
}
\description{
This function plots the original specification curve on top of
95\% quantiles of the bootstrapped 'under-the-null' distributions.
}
