% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{plot.specr.object}
\alias{plot.specr.object}
\title{Plot specification curve and analytic choices}
\usage{
\method{plot}{specr.object}(
  x,
  type = "default",
  var = estimate,
  choices = c("x", "y", "model", "controls", "subsets"),
  labels = c("A", "B"),
  rel_heights = c(2, 3),
  desc = FALSE,
  null = 0,
  ci = TRUE,
  ribbon = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{specr.object} object, usually resulting from calling \code{specr()}.}

\item{type}{What type of figure should be plotted? If \code{what = "default"},
the standard specification curve analysis plot (the specification curve as the
upper panel and an overview of the relevant choices as the lower panel) is
created. If \code{what = "curve"}, only the specification curve (upper panel
of the default plot) is plotted. If \code{what = "choices"} only the choice
panel (lower part of the default plot) is plotted. \code{what = "boxplot"},
an alternative visualization of differences between choices is plotted. See
examples for more information.}

\item{var}{Which parameter should be plotted in the curve? Defaults to
\code{estimate}, but other parameters (e.g., p.value, fit_r.squared,...)
can be plotted too.}

\item{choices}{A vector specifying which analytic choices should be plotted.
By default, all choices (x, y, model, controls, subsets) are plotted.}

\item{labels}{Labels for the two parts of the plot}

\item{rel_heights}{vector indicating the relative heights of the plot.}

\item{desc}{Logical value indicating whether the curve should the arranged in
a descending order. Defaults to FALSE.}

\item{null}{Indicate what value represents the 'null' hypothesis (defaults to
zero).}

\item{ci}{Logical value indicating whether confidence intervals should be
plotted.}

\item{ribbon}{Logical value indicating whether a ribbon instead should be
plotted.}

\item{...}{additional arguments that can be passed to \code{plot_grid()}.}
}
\value{
A \link[ggplot2]{ggplot} object that can be customized further.
}
\description{
This function plots visualizations of the specification curve
analysis. The function requires an object of class \code{specr.object}, usually
the results of calling \code{specr()} to create a standard visualization of the
specification curve analysis. Several types of visualizations are possible.
}
\examples{
# Setup specifications
specs <- setup(data = example_data,
               y = c("y1", "y2"),
               x = c("x1", "x2"),
               model = "lm",
               distinct(example_data, group1),
               distinct(example_data, group2),
               controls = c("c1", "c2"))

# Run analysis
results <- specr(specs, workers = 1)

# Plot results in various ways
plot(results)
plot(results, choices = c("x", "y"))
plot(results, ci = FALSE, ribbon = TRUE)
plot(results, type = "curve", desc = TRUE)
plot(results, type = "boxplot") + scale_fill_brewer(palette = "Dark2")

# Also other variables in the resulting data set can be plotted
plot(results, type = "curve", var = fit_r.squared, ci = FALSE)

# Such a plot can also be extended (e.g., by again adding the estimates with
# confidence intervals)
plot(results, type = "curve", var = fit_r.squared) +
  geom_point(aes(y = estimate), shape = 5) +
  labs (x = "specifications", y = "r-squared | estimate")

}
