% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{plot.specr}
\alias{plot.specr}
\title{Plot specification curve and analytic choices}
\usage{
\method{plot}{specr}(
  x,
  what = "default",
  choices = c("x", "y", "model", "controls", "subsets"),
  labels = c("A", "B"),
  rel_heights = c(2, 3),
  desc = FALSE,
  null = 0,
  ci = TRUE,
  ribbon = FALSE,
  ...
)
}
\arguments{
\item{x}{a "specr" object, usually resulting from calling \code{specr()}.}

\item{choices}{a vector specifying which analytical choices should be plotted. By default, all choices are plotted.}

\item{labels}{labels for the two parts of the plot}

\item{rel_heights}{vector indicating the relative heights of the plot.}

\item{desc}{logical value indicating whether the curve should the arranged in
a descending order. Defaults to FALSE.}

\item{null}{Indicate what value represents the 'null' hypothesis (defaults to
zero).}

\item{ci}{logical value indicating whether confidence intervals should be
plotted.}

\item{ribbon}{logical value indicating whether a ribbon instead should be
plotted.}

\item{...}{additional arguments that can be passed to \code{plot_grid()}.}
}
\value{
a \link[ggplot2]{ggplot} object.
}
\description{
This function plots an entire visualization of the specification curve
analysis. The function uses the "specr" object that is produced by \code{specr()} to creates a standard visualization of the specification curve analysis.
}
\examples{
# Setup specifications
specs <- setup(data = example_data,
               y = c("y1", "y2"),
               x = c("x1", "x2"),
               model = "lm",
               distinct(example_data, group1),
               distinct(example_data, group2),
               controls = c("c1", "c2"))

# Run analysis
results <- specr(specs, workers = 1)

# Plot results in various ways
plot(results)
plot(results, choices = c("x", "y"))
plot(results, ci = F, ribbon = T)
plot(results, what = "curve", desc = T)
plot(results, what = "boxplot")
}
