% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curve.r
\name{plot_curve}
\alias{plot_curve}
\title{Plot specification curve}
\usage{
plot_curve(df, desc = FALSE, ci = TRUE, prob = 0.975,
  legend = FALSE)
}
\arguments{
\item{df}{a data frame containing the choices and results of each specification (resulting from \code{run_specs}).}

\item{desc}{logical value indicating whether the curve should the arranged in a descending order. Defaults to FALSE.}

\item{ci}{logical value indicating whether confidence intervals should be plotted.}

\item{prob}{numeric value indicating what type of confidence intervals should be plotted. Defaults to .975 (= 95% confidence intervalls.)}

\item{legend}{logical value indicating whether the legend should be plotted Defaults to FALSE.}
}
\value{
a ggplot object.
}
\description{
This function plots the a ranked specification curve. Confidence intervals can be included. Significant specifications are highlighted. Further customization with ggplot2 is possible.
}
