% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decision.r
\name{plot_decisions}
\alias{plot_decisions}
\title{Plot the decision tree}
\usage{
plot_decisions(df, label = FALSE, legend = FALSE)
}
\arguments{
\item{df}{data frame resulting from \code{run_specs()}.}

\item{label}{Logical. Should labels be included? Defaults to FALSE. Produces only a reasonable plot if number of specifications is low.}

\item{legend}{Logical. Should specific decisions be identifiable. Defaults to FALSE.}
}
\description{
This function plots a simple decision tree that can be used to understand how analytical choices produce a certain number of specifications. It is somewhat useless if the number of specifications is very high.
}
\examples{
results <- run_specs(df = example_data,
                    y = c("y1", "y2"),
                    x = c("x1", "x2"),
                    model = c("lm"),
                    controls = c("c1", "c2"))

# Basic, non-labelled decisions tree
plot_decisions(results)

# Labelled decisions tree
plot_decisions(results, label = TRUE)

# Identify choices
plot_decisions(results, legend = TRUE)

# All combined
plot_decisions(results, label = TRUE, legend = TRUE)
}
