% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_samplesizes.r
\name{plot_samplesizes}
\alias{plot_samplesizes}
\title{Plot sample sizes}
\usage{
plot_samplesizes(df, desc = FALSE)
}
\arguments{
\item{df}{a data frame containing the choices and results of each specification (resulting from \code{run_specs}).}

\item{desc}{logical value indicating whether the curve should the arranged in a descending order. Defaults to FALSE.}
}
\description{
Plot sample sizes
}
