% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.r
\name{plot_summary}
\alias{plot_summary}
\title{A boxplot summarizing statistics per analytical choice}
\usage{
plot_summary(df, choices = c("x", "y", "model", "controls", "subsets"))
}
\arguments{
\item{df}{a data frame containing the choices and results of each specification (resulting from \code{run_specs}).}

\item{choices}{a vector specifying which analytical choices should be plotted. By default, all choices are plotted.}
}
\value{

}
\description{
This function provides a convenient way to investigate the influence of individual choices on the estimate of interest. Similar to \code{summarise_specs}), it can be used to investigate individual choices.
}
\examples{
# run spec analysis
results <- run_specs(example_data,
                     y = c("y1", "y2"),
                     x = c("x1", "x2"),
                     model = "lm",
                     controls = c("c1", "c2"),
                     subset = list(group1 = unique(example_data$group1)))
plot_summary(results, choices = c("subsets", "controls", "y"))
}
