% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSpec.R
\name{plotSpec}
\alias{plotSpec}
\title{A Spectrum Plotter}
\usage{
plotSpec(path = NA)
}
\arguments{
\item{data}{Data to plot.}
}
\description{
This function allows you to plot spectra.
}
\examples{
plotSpec(data)
}
\keyword{plot}

