% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{print.specr.boot}
\alias{print.specr.boot}
\title{Print method for S3 class "specr.book"}
\usage{
\method{print}{specr.boot}(x, ...)
}
\description{
Print method for S3 class "specr.book"
}
\keyword{internal}
