% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{print.specr.object}
\alias{print.specr.object}
\title{Print method for S3 class "specr.object"}
\usage{
\method{print}{specr.object}(x, ...)
}
\description{
Reports the number of successful and failed specifications and, when
available, prints a descriptive summary of the successful results.
}
\keyword{internal}
