% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{print.specr.setup}
\alias{print.specr.setup}
\title{Print method for S3 class "specr.setup"}
\usage{
\method{print}{specr.setup}(x, ...)
}
\description{
Print method for S3 class "specr.setup"
}
\keyword{internal}
