% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_specs.r
\name{run_specs}
\alias{run_specs}
\title{Run specification curve analysis}
\usage{
run_specs(df, y, x, model, controls, subsets)
}
\arguments{
\item{subsets}{}
}
\description{
Run specification curve analysis
}
