% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_specs.r
\name{setup_specs}
\alias{setup_specs}
\title{Create result frame with specifications}
\usage{
setup_specs(independent, dependent, control = NA, subset = NA,
  transform = NA)
}
\arguments{
\item{independent}{Enter independent variable as vector.}

\item{transform}{}
}
\description{
Create result frame with specifications
}
\examples{
# Create variable for subsetting
gender <- c(rep(c("male", "female"), 10))

# Setup results frame with specifications
setup_specs(independent = c("tv_use", "smartphone_use"),
            dependent = c("depression", "loneliness"),
            control = c("self_esteem", "number_friends"),
            subset = gender)


}
