% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_specs.r
\name{summarise_specs}
\alias{summarise_specs}
\title{Summarise the specification curve results}
\usage{
summarise_specs(df, group = NULL)
}
\arguments{
\item{df}{a data frame containing the choices and results of each specification (resulting from \code{run_specs}).}

\item{group}{a grouping factor (e.g., "subsets"). Defaults to NULL.}
}
\description{
This function allows to inspect results of the specification curves by returning a comparatively simple summary of the results. These results can be returned for specific analytical choices.
}
