% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{summary.specr.boot}
\alias{summary.specr.boot}
\title{Summarizing the bootstrapped results}
\usage{
\method{summary}{specr.boot}(x, group = NULL, ...)
}
\arguments{
\item{x}{A \code{specr.boot} object, resulting from calling \code{boot_null()}.}

\item{group}{Variables indicating which variables to summarize the results by}

\item{...}{further arguments passed to or from other methods (currently ignored).}
}
\value{
A tibble.
}
\description{
Generic summary function for an object of class \code{specr.boot} (resulting from
\code{boot_null()}. Provides an approach to inference for specification curve analysis
consists of obtaining the median effect estimated across all specifications,
and then testing whether this median estimated effect is more extreme than
would be expected if all specifications had a true effect of zero.
}
