% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{summary.specr}
\alias{summary.specr}
\title{Summarizing the Specification Curve Analysis}
\usage{
\method{summary}{specr}(
  x,
  what = "default",
  ...,
  stats = list(median = median, mad = mad, min = min, max = max, q25 = function(x)
    quantile(x, prob = 0.25), q75 = function(x) quantile(x, prob = 0.75)),
  digits = 2,
  rows = 6
)
}
\arguments{
\item{x}{an object of class "specr", usually resulting of a call to \code{specr}.}

\item{what}{different aspects can be summarized and printed. See details for alternative summaries}

\item{...}{In combination with \code{what = "curve"}, provide one or more variables (e.g., subsets, controls,...) that denote the available analytic choices to group summary of the estimate.}

\item{stats}{named vector or named list of summary functions (individually defined summary functions can included). If it is not named, placeholders (e.g., "fn1") will be used as column names.}

\item{digits}{the number of digits to use when printing the specification table.}

\item{rows}{the number of rows of the specification tibble that should be printed.}
}
\description{
\code{summary} method for class "specr".
}
\examples{
# Setup up specifications (returns "specr.setup" object)
specs <- setup(data = example_data,
               y = c("y1", "y2"),
               x = c("x1", "x2"),
               model = "lm",
               distinct(example_data, group1),
               controls = c("c1", "c2"))

# Run analysis (returns "specr" object)
results <- specr(specs, workers = 1)

# Summary of the "specr" object
summary(results)  # Default
summary(results, digits = 5, rows = 10)
summary(results, what = "curve")
summary(results, what = "curve", x, y)
summary(results, what = "curve",
       x, group1,
       stats = list(median = median,
                    min = min,
                    max = max))
}
\seealso{
The function used to create the "specr.setup" object: \code{setup}.
}
