% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{summary.specr.setup}
\alias{summary.specr.setup}
\title{Summarizing the Specifications Setup}
\usage{
\method{summary}{specr.setup}(x, digits = 2, rows = 6, print.specs = FALSE)
}
\arguments{
\item{x}{an object of class "specr.setup", usually, a result of a call to \code{setup}.}

\item{digits}{the number of digits to use when printing the specification table.}

\item{rows}{the number of rows of the specification tibble that should be printed.}

\item{print.specs}{logical; if \code{TRUE}, a head of the specification tibble is returned and printed.}
}
\description{
\code{summary} method for class "specr.setup".
}
\examples{
# Setup specifications
specs <- setup(data = example_data,    # data
               x = c("x1", "x2"),      # independent variables
               y = c("y1"),            # dependent variable
               model = c("lm", "glm"), # model functions
               controls = "c1")        # covariates to be added
summary(specs)
summary(specs, print.specs = TRUE, rows = 10)
}
\seealso{
The function used to create the "specr.setup" object, \code{setup}.
}
