% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_specs.r
\name{variance_specs}
\alias{variance_specs}
\title{Get variance component of the specification curve}
\usage{
variance_specs(df, estimate, percent = TRUE)
}
\arguments{
\item{df}{a data frame containing the choices and results of each specification (resulting from \code{run_specs}).}

\item{estimate}{the statistic to be evaluated (e.g., the unstandardized effect or the p-value)}

\item{percent}{a logical value indicating whether the icc should also be printed as percentage.}
}
\description{
Get variance component of the specification curve
}
