% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_null.r
\name{boot_null}
\alias{boot_null}
\title{Inference with specification curve analysis}
\usage{
boot_null(x, y, n_samples = 500)
}
\arguments{
\item{x}{A fitted \code{specr.object}, resulting from running \code{specr()}.}

\item{y}{A \code{specr.setup} object, resulting from running \code{setup()}}

\item{n_samples}{Number of bootstrap samples to be drawn. Defaults to 500.}
}
\value{
a object of class \code{specr.boot} that represents a list with the original and bootstrapped results.
}
\description{
It is possible to generate the distributions for any of these test statistics under-the-null analytically (that is, with statistical formulas), because the specifications are neither statistically independent nor part of a single model.
This function hence generates such distributions by relying on resampling under-the-null. This involves modifying the observed data so that the null hypothesis is known to be true, and then drawing random samples of the modified data.
The resulting object of class \code{specr.boot} can be summarized and plotted with generic functions.
}
\examples{
# Setup up  specifications

# Requires to keep full model
tidy_full <- function(x) {
fit <- broom::tidy(x, conf.int = TRUE)
fit$res <- list(x)  # Store model object
return(fit)
}

specs <- setup(data = example_data,
   y = c("y1", "y2"),
   x = c("x1", "x2"),
   model = "lm",
   controls = c("c1", "c2"),
   fun1 = tidy_full)

# Run analysis
results <- specr(specs)

# Run bootstrapping
boot_models <- boot_null(results, specs, n_samples = 10) # better 1,000!

# Summarize findings
summary(boot_models)

# Plot under-the-null curves on top of specification curve
plot(boot_models)

}
\references{
\itemize{
\item Simonsohn, U., Simmons, J.P. & Nelson, L.D. (2020). Specification curve analysis. \emph{Nature Human Behaviour, 4}, 1208–1214. https://doi.org/10.1038/s41562-020-0912-z
}
}
\keyword{internal}
