% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.4.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data set}
\format{
A tibble
}
\usage{
data(example_data)
}
\description{
This simulated data set can be used to explore the major function of 'specr'.
}
\examples{
data(example_data)
head(example_data)
}
\keyword{datasets}
