% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_specs.r
\name{icc_specs}
\alias{icc_specs}
\title{Get variance component of the specification curve}
\usage{
icc_specs(model, percent = TRUE)
}
\arguments{
\item{model}{a multilevel model that captures the variances of the specification curve (resulting from \code{run_specs}).}

\item{percent}{a logical value indicating whether the icc should also be printed as percentage.}
}
\value{

}
\description{
Get variance component of the specification curve
}
\examples{
# Step 1: Run spec curve analysis
results <- run_specs(df = example_data,
                     y = c("y1", "y2"),
                     x = c("x1", "x2"),
                     model = c("lm"))

# Step 2: Estimate multilevel model
model <- lme4::lmer(estimate ~ 1 + (1|x)  + (1|y), data = results)

# Step 3: Estimate intra-class correlation
icc_specs(model)
}
