% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_choices.r
\name{plot_choices}
\alias{plot_choices}
\title{Plot how analytical choices affect results}
\usage{
plot_choices(df, desc = FALSE)
}
\arguments{
\item{df}{a data frame containing the choices and results of each specification (resulting from \code{run_specs}).}

\item{desc}{logical value indicating whether the curve should the arranged in a descending order. Defaults to FALSE.}
}
\value{
a ggplot object.
}
\description{
This functions plots how the analytical choices affect the obtained results (i.e., the rank within the curve). Significant results are highlighted. Further customization via ggplot2 is possible.
}
