% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specs.r
\name{plot_specs}
\alias{plot_specs}
\title{Plot specification curve}
\usage{
plot_specs(df, ci = TRUE)
}
\arguments{
\item{df}{data frame resulting from run_specs().}

\item{ci}{logical value indicating whether confidence intervals should be plotted.}
}
\description{
Plot specification curve
}
