% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variance.r
\name{plot_variance}
\alias{plot_variance}
\title{Plot variance components of the specification curve}
\usage{
plot_variance(df, palette = "Set3")
}
\arguments{
\item{df}{a data frame containing the choices and results of each specification (resulting from \code{run_specs}).}

\item{palette}{see "palette" within ggplot2. Defaults to "Set3".}
}
\description{
Plot variance components of the specification curve
}
