% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Specifying analytical decisions in a specification setup}
\usage{
setup(
  data,
  x,
  y,
  model,
  ...,
  controls = NULL,
  add_to_formula = NULL,
  fun = function(x) broom::tidy(x, conf.int = TRUE)
)
}
\arguments{
\item{data}{The data set that should be used for the analysis}

\item{x}{A vector denoting independent variables}

\item{y}{A vector denoting the dependent variables}

\item{model}{A vector denoting the model(s) that should be estimated.}

\item{...}{Specification of potential subsets/groups. As the subsetting
variable should be in the data set, please specify as follows
(\code{distinct(data, variable)}). Note: This variable needs to be a factor in
the data set! See also examples further below.}

\item{controls}{A vector of the control variables that should be included.
Defaults to NULL.}

\item{add_to_formula}{A string specifying aspects that should always be
included in the formula (e.g. a constant covariate, random effect structures...)}

\item{fun}{A function that extracts the parameters of interest from the
fitted models. Defaults to \link[broom:broom]{tidy}, which works
with a large range of different models.}
}
\value{
An object of class \code{specr.setup} which includes all possible
specifications based on combinations of the analytic choices. The
resulting list includes a specification tibble, the data set, and additional
information about the universe of specifications. Use
\code{methods(class = "specr.setup")} for an overview on available methods.
}
\description{
Creates all possible specifications as a combination of
different dependent and independent variables, model types, control
variables, potential subset analyses, as well as potentially other
analytic choices. This function represents the first step in the
analytic framework implemented in the package \code{specr}. The resulting
class \code{specr.setup} then needs to be passed to the core function of
the package called \code{specr()}, which fits the specified models across
all specifications.
}
\details{
Empirical results are often contingent on analytical decisions that
are equally defensible, often arbitrary, and motivated by different reasons.
This decisions may introduce bias or at least variability. To this end,
specification curve analyses  (Simonsohn et al., 2019) or multiverse
analyses (Steegen et al., 2016) refer to identifying the set of
theoretically justified, statistically valid (and potentially also non-redundant
specifications, fitting the "multiverse" of models represented by these
specifications and estract relevant parameters often to display the results
graphically as a so-called specification curve. This allows readers to
identify consequential specifications decisions and how they affect the results
or parameter of interest.

\bold{Use of this function}

A general overview is provided in the vignettes \code{vignette("specr")}.
It is assumed that you want to estimate the relationship between two variables
(\code{x} and \code{y}). What varies may be what variables should be used for
\code{x} and \code{y}, what model should be used to estimate the relationship,
and whether the relationship should be estimated for certain subsets. This
allows to (re)produce almost any analytical decision imaginable. See examples
below for how a number of typical analytical decision can be implemented.
Afterwards you pass the resulting object of a class \code{specr.setup} to the
function \code{specr()} to run the specification curve analysis.

Note, the resulting class of \code{specr.setup} allows to use generic functions.
Use \code{methods(class = "specr.setup")} for an overview on available methods and
e.g., \code{?summary.specr.setup} to view the dedicated help page.
}
\examples{
# Example 1 ----
# Setting up typical specifications
specs <- setup(data = example_data,
               x = c("x1", "x2"),
               y = c("y1", "y2"),
               model = "lm",
               distinct(example_data, group1),
               distinct(example_data, group2),
               controls = c("c1", "c2"))

# Check specifications
summary(specs)


# Example 2 ----
# Setting up specifications for multilevel models
specs <- setup(data = example_data,
               x = c("x1", "x2"),
               y = c("y1", "y2"),
               model = c("lmer"),              # multilevel model
               distinct(example_data, group1),
               controls = c("c1", "c2"),
               add_to_formula = "(1|group2)")  # include random effect in all formulas

# Check specifications
summary(specs)


# Example 3 ----
# Setting up specifications with specific additions to the formula and different parameter extract functions

# Create custom extract function
tidy_99 <- function(x) {
  broom::tidy(x, conf.int = TRUE, conf.level = .99)
}

# Setup specs
specs <- setup(data = example_data,
               x = c("x1", "x2"),
               y = c("y1", "y2"),
               model = "lm",
               fun = tidy_99,               # pass new function to setup
               add_to_formula = "c1 + c2")  # add covariates to all models

# Check specifications
summary(specs)


# Example 4 ----
# Specific specification that can be represent as "subsets"
library(tidyverse)
# Create variable that defines outliers
dat <- mutate(example_data, outlier = ifelse(x1 > 2 | x1 < -2, "outlier", "no outlier"))
table(dat$outlier)

# Setup specs
specs <- setup(data = dat,
               x = c("x1", "x2"),
               y = c("y1", "y2"),
               model = "lm",
               distinct(d, outlier))

# Remove outlier subset from resulting data set to avoid non-meaningful specifications
specs$specs <- filter(specs$specs, subsets != "outlier")

# Check specifications
summary(specs)
}
\references{
\itemize{
\item Simonsohn, U., Simmons, J. P., & Nelson, L. D. (2019). Specification Curve: Descriptive and Inferential Statistics for all Plausible Specifications. Available at: https://doi.org/10.2139/ssrn.2694998
\item Steegen, S., Tuerlinckx, F., Gelman, A., & Vanpaemel, W. (2016). Increasing Transparency Through a Multiverse Analysis. Perspectives on Psychological Science, 11(5), 702-712. https://doi.org/10.1177/1745691616658637
}
}
\seealso{
\code{\link[=specr]{specr()}} to run the actual specification curve analysis
}
