% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_specs.r
\name{setup_specs}
\alias{setup_specs}
\title{Create result frame with specifications}
\usage{
setup_specs(x, y, model = "glm", controls = NA)
}
\arguments{
\item{x}{Enter x variable as vector.}

\item{controls}{}
}
\description{
Create result frame with specifications
}
\examples{
# Setup results frame with specifications
setup_specs(x = c("tv_use", "smartphone_use"),
            y = c("depression", "loneliness"),
            controls = c("self_esteem", "number_friends"))


}
