% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_specs.r
\name{setup_specs}
\alias{setup_specs}
\title{Set up specifications}
\usage{
setup_specs(x, y, model, controls = NULL)
}
\arguments{
\item{x}{a vector of the independent variables}

\item{y}{a vector of the dependent variables}

\item{model}{a vector of the type of models that should be estimated}

\item{controls}{a vector of the control variables that should be included. Defaults to none.}
}
\value{
a \link[tibble:tibble-package]{tibble} that includes all combinations of determined analytical choices.
}
\description{
This function creates a tibble that includes all possible specifications based on different dependent and independent variables, different types of models, and different control variables. This function can be used to check the combinations of analytical choices. Otherwise, the function is called within \code{\link[=run_specs]{run_specs()}}.
}
\examples{
setup_specs(y = c("y1"),
            x = c("x1", "x2"),
            model = c("lm"),
            controls = c("c1", "c2"))

}
\seealso{
\code{\link[=run_specs]{run_specs()}} to run the specification curve analysis.
}
