% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_specs.r
\name{setup_specs}
\alias{setup_specs}
\title{Set up specifications}
\usage{
setup_specs(x, y, model, controls = NULL, all.comb = FALSE)
}
\arguments{
\item{x}{a vector denoting independent variables}

\item{y}{a vector denoting the dependent variables}

\item{model}{a vector denoting the model(s) that should be estimated.}

\item{controls}{a vector of the control variables that should be included. Defaults to NULL.}

\item{all.comb}{a logical value indicating what type of combinations of the control variables should be specified. Defaults to FALSE (i.e., none, all, and each individually). If this argument is set to TRUE, all possible combinations between the control variables are specified (see examples).}
}
\value{
a \link[tibble:tibble-package]{tibble} that includes all possible specifications based on combinations of the analytic choices.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated because the new version of specr uses a new analytic framework. In this framework, you should use the function \code{\link[=setup]{setup()}} instead.
This function creates a tibble that includes all possible specifications based the dependent and independent variables, model types, and control variables that are specified. This function simply produces a tibble of all combinations. It can be used to check the specified analytical choices. This function is called within \code{\link[=run_specs]{run_specs()}}, which estimates all specified models based on the data that are provided.
}
\examples{
setup_specs(x = c("x1", "x2"),
            y = "y2",
            model = "lm",
            controls = c("c1", "c2"))

}
\seealso{
\code{\link[=run_specs]{run_specs()}} to run the specification curve analysis.
}
\keyword{internal}
