% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specr.R
\name{specr}
\alias{specr}
\title{Estimate all specifications}
\usage{
specr(specs, data = NULL, workers = availableCores(), message = TRUE)
}
\arguments{
\item{specs}{a \code{specr.setup} object resulting from \code{setup()} or a tibble that contains the relevant specifications (e.g., a tibble resulting from \code{setup()})}

\item{data}{a data frame or tibble that contains the data set. Does not have to be specified if specs are a \code{specr.setup} object.}

\item{workers}{A numeric value that indicates how many cores should be used. Defaults to \code{availableCores()} from the future package.}
}
\value{
a \code{specr} object that includes all specifications and the results.
}
\description{
This is the central function of the package. It
}
\details{
It takes the data frame and vectors for analytical choices related to the dependent variable, the independent variable, the type of models that should be estimated,the set of covariates that should be included (none, each individually, and all together),
as well as a named list of potential subsets.
}
\examples{
# Example 1 ----
# Setup up specifications (returns "specr.setup" object)
specs <- setup(data = example_data,
               y = c("y1", "y2"),
               x = c("x1", "x2"),
               model = "lm",
               distinct(example_data, group1),
               controls = c("c1", "c2"))

# Run analysis (returns "specr" object)
results <- specr(specs,
                 workers = 1)   # Not parallelized, using simple `map` functions

# Summary of the "specr" object
summary(results)
summary(results, what = "curve", subsets)

# Example 2 ----
# Setup up specifications, but return tibble (if you don't want to work with S3 classes)
specs2 <- setup(data = example_data,
                y = c("y1", "y2"),
                x = c("x1", "x2"),
                model = "lm",
                distinct(example_data, group1),
                distinct(example_data, group2),
                controls = "c1",
                add_to_formula = "c2",
                to_tibble = TRUE)               # set to TRUE to return tibble

# Run analysis, but add data as "specs" is only a tibble of the specifications and doesn't include the data
results2 <- specr(specs2,
                  data = example_data, # provide data
                  workers = 4)         # using 4 cores and `future_map` functions

# Check results (returned as a tibble in this case)
summary(results2)

# Get tibble from "specr" object
as_tibble(results2)
}
\references{
\itemize{
\item Simonsohn, U., Simmons, J. P., & Nelson, L. D. (2019). Specification Curve: Descriptive and Inferential Statistics for all Plausible Specifications. Available at: https://doi.org/10.2139/ssrn.2694998
\item Steegen, S., Tuerlinckx, F., Gelman, A., & Vanpaemel, W. (2016). Increasing Transparency Through a Multiverse Analysis. Perspectives on Psychological Science, 11(5), 702-712. https://doi.org/10.1177/1745691616658637
}
}
