% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_specs.r
\name{summarise_specs}
\alias{summarise_specs}
\title{Summarise the specification curve results}
\usage{
summarise_specs(df, var = estimate, stats = lst(median, mad, min, max,
  q25 = function(x) quantile(x, prob = 0.25), q75 = function(x) quantile(x,
  prob = 0.75)), group = NULL)
}
\arguments{
\item{df}{a data frame containing the choices and results of each specification (resulting from \code{run_specs}).}

\item{var}{which variable should be evaluated? Defaults to estimate (the effect sizes computed by \code{run_specs}). No need to use closures.}

\item{stats}{list object (\code{lst()}) of summary functions.}

\item{group}{a grouping factor (e.g., "subsets"). Several grouping variables can be passed. Defaults to NULL.}
}
\description{
This function allows to inspect results of the specification curves by returning a comparatively simple summary of the results. These results can be returned for specific analytical choices.
}
