% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_specs.r
\name{summarise_specs}
\alias{summarise_specs}
\title{Summarise the specification curve results}
\usage{
summarise_specs(df, var = estimate, stats = lst(median, mad, min, max,
  q25 = function(x) quantile(x, prob = 0.25), q75 = function(x) quantile(x,
  prob = 0.75)), group = NULL)
}
\arguments{
\item{df}{a data frame containing the choices and results of each specification (resulting from \code{run_specs}).}

\item{var}{which variable should be evaluated? Defaults to estimate (the effect sizes computed by \code{run_specs}). No need to use closures.}

\item{stats}{list object (\code{lst()}) of summary functions.}

\item{group}{a grouping factor (e.g., "subsets"). Several grouping variables can be passed. Defaults to NULL.}
}
\description{
This function allows to inspect results of the specification curves by returning a comparatively simple summary of the results. These results can be returned for specific analytical choices.
}
\examples{
# Run specification curve analysis
results <- run_specs(df = example_data,
                     y = c("y1", "y2"),
                     x = c("x1", "x2"),
                     model = c("lm", "lm_gauss"),
                     controls = c("c1", "c2"),
                     subsets = list(group1 = unique(example_data$group1),
                                    group2 = unique(example_data$group2)))

# Basic example
summarise_specs(results)

# Summary of specific analytical choices
summarise_specs(results, group = c("subsets", "x"))

# Summare of other estimates
summarise_specs(results, var = "p.value", group = "controls")
}
